Correct printf() format specifiers for size_t values.
"%u" can't be used to print size_t values under 64 bit architectures where
sizeof(size_t) != sizeof(int). Even though it seems to work somehow with glibc
under amd64 (possibly because it aligns all parameters on 64 bit boundary?),
it is still wrong so use "%lu" and cast size_t to unsigned long explicitly.

+++ src/gui/wxMDialogs.cpp
@@ -2878,7 +2878,8 @@ void MProgressInfo::SetLabel(const wxString& label)
 
 void MProgressInfo::SetValue(size_t numDone)
 {
-   SetLabel(wxString::Format(_("%u done"), numDone));
+   SetLabel(wxString::Format(_("%lu done"),
+                             static_cast<unsigned long>(numDone)));
 }
 
 MProgressInfo::~MProgressInfo()
