+++ src/gui/wxMIMETreeDialog.cpp
@@ -225,7 +225,8 @@ wxMIMETreeDialog::wxMIMETreeDialog(const MimePart *partRoot,
 
    m_treectrl->Expand(m_treectrl->GetRootItem());
 
-   box->SetLabel(wxString::Format(_("%u MIME parts"), m_countParts));
+   box->SetLabel(wxString::Format(_("%lu MIME parts"),
+                                  static_cast<unsigned long>(m_countParts)));
 
 
    SetSizer(sizerTop);
